refactor: rename stores/ to graph/
Pure directory and include-path rename. Prepares the codebase for the
introduction of TemporalGraphData and related RAII data types in the
next commit. No semantic change.
