+++ temporal_random_walk/src/graph/edge_data.cuh
