+++ temporal_random_walk/src/graph/node_features.cuh
