+++ temporal_random_walk/src/graph/temporal_graph.cuh
