+++ temporal_random_walk/src/proxies/EdgeData.cuh
@@ -2,7 +2,7 @@
 #define EDGE_DATA_H
 
 #include <vector>
-#include "../stores/edge_data.cuh"
+#include "../graph/edge_data.cuh"
 #include "../common/error_handlers.cuh"
 
 #ifdef HAS_CUDA
